bet tests

+++ app/Http/Controllers/BetController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreBetRequest;
@@ -6,24 +7,20 @@ use App\Http\Requests\UpdateBetRequest;
 use App\Http\Resources\BetResource;
 use App\Models\Bet;
 use App\Repositories\BetRepository;
-use App\Repositories\CompetitionRepository;
 use Illuminate\Http\Request;
 
 class BetController extends Controller
 {
     private $betRepository;
 
-    private $competitionRepository;
-
     private $bettorService;
 
     public function __construct(
         BetRepository $betRepository,
-        CompetitionRepository $competitionRepository,
         // BettorService $bettorService
     ) {
         $this->betRepository = $betRepository;
-        $this->competitionRepository = $competitionRepository;
+
         // $this->bettorService = $bettorService;
     }
 
@@ -37,42 +34,43 @@ class BetController extends Controller
 
     public function store(StoreBetRequest $request)
     {
-        // $this->authorize('create', Bet::class);
         try {
             $bet = $this->betRepository->store($request->all());
 
             return new BetResource($bet);
         } catch (\Throwable $th) {
-            return response()->json(['error' => 'Impossible to create the bet.' . $th->getMessage()]);
+            return response()->json(['error' => 'Impossible to create the bet.'.$th->getMessage()]);
         }
     }
 
     public function show(Bet $bet)
     {
+        $this->authorize('view', $bet);
+
         return new BetResource($bet);
     }
 
     public function update(UpdateBetRequest $request, Bet $bet)
     {
-        // $this->authorize('update', $bet);
+        // see the autjorization in the UpdateBetRequest
         try {
             $bet = $this->betRepository->update($bet, $request->all());
 
             return new BetResource($bet);
         } catch (\Throwable $th) {
-            return response()->json(['error' => 'Impossible to update the bet.' . $th->getMessage()]);
+            return response()->json(['error' => 'Impossible to update the bet.'.$th->getMessage()]);
         }
     }
 
     public function destroy(Bet $bet)
     {
-        // $this->authorize('delete', $bet);
+        $this->authorize('delete', $bet);
         try {
             $bet = $this->betRepository->destroy($bet);
 
             return response()->noContent();
         } catch (\Throwable $th) {
-            return response()->json(['error' => 'Impossible to delete the bet.' . $th->getMessage()]);
+            return response()->json(['error' => 'Impossible to delete the bet.'.$th->getMessage()]);
         }
     }
 }
